+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added Pick from image from gallery
 * Added `photos` icon to the icons library
-* Support dor `onSelectedTabTouch` event
+* Support for `onSelectedTabTouch` event
 
 ### Changed
 
@@ -38,3 +38,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Support for attachments with names containing white spaces
+* Height of Textarea component in ios